make bsp structures and file formats use float instead of double

+++ em-game/shared/bsp.h
@@ -1,9 +1,9 @@
 
 struct bspnode_t
 {
-	double x1, y1, x2, y2;
-	double tstart, tend;
-	double dtstart, dtend;
+	float x1, y1, x2, y2;
+	float tstart, tend;
+	float dtstart, dtend;
 
 	struct bspnode_t *front, *back;
 };
@@ -11,8 +11,8 @@ struct bspnode_t
 #ifdef _ZLIB_H
 int load_bsp_tree(gzFile file);
 #endif
-struct bspnode_t *circle_walk_bsp_tree(double xdis, double ydis, double r, double *t_out);
-struct bspnode_t *line_walk_bsp_tree(double x1, double y1, double x2, double y2);
+struct bspnode_t *circle_walk_bsp_tree(float xdis, float ydis, float r, float *t_out);
+struct bspnode_t *line_walk_bsp_tree(float x1, float y1, float x2, float y2);
 void rail_walk_bsp(float x1, float y1, float x2, float y2, float *out_x, float *out_y);
 
 extern struct bspnode_t *bspnode0;
